+++ CODE/PY_FILES/Phase-1/OUTPUTS/ckt_details_c17.txt
@@ -1,0 +1,31 @@
++-----------------+---------+
+| Type            |   Count |
++=================+=========+
+| Primary Inputs  |       5 |
++-----------------+---------+
+| Primary Outputs |       2 |
++-----------------+---------+
+| INPUT gates     |       5 |
++-----------------+---------+
+| NAND gates      |       6 |
++-----------------+---------+
+
+
+----------------------------------------------------------------------
+	 	 FAN-OUT
+----------------------------------------------------------------------
+NAND-10: NAND-22
+NAND-11: NAND-16, NAND-19
+NAND-16: NAND-22, NAND-23
+NAND-19: NAND-23
+NAND-22: OUTPUT-22
+NAND-23: OUTPUT-23
+----------------------------------------------------------------------
+	 	 FAN-IN
+----------------------------------------------------------------------
+NAND-10 : INPUT-1, INPUT-3
+NAND-11 : INPUT-3, INPUT-6
+NAND-16 : INPUT-2, NAND-11
+NAND-19 : NAND-11, INPUT-7
+NAND-22 : NAND-10, NAND-16
+NAND-23 : NAND-16, NAND-19
