+++ CODE/PY_FILES/Phase-2/OUTPUTS/ckt_traversal_c17.txt
@@ -1,0 +1,25 @@
+
+Circuit Delay  :  64.68834289793794 ps
+
+----------------------------------------------------------------------
+	 	 GATE SLACKS
+----------------------------------------------------------------------
+INPUT-1 : 31.629008569012157 ps
+INPUT-2 : 24.339789909919176 ps
+INPUT-3 : 5.880758445267084 ps
+INPUT-6 : 5.880758445267084 ps
+INPUT-7 : 24.237080755551016 ps
+NAND-10 : 31.629008569012157 ps
+NAND-11 : 5.880758445267084 ps
+NAND-16 : 5.880758445267087 ps
+NAND-19 : 11.072632556188681 ps
+NAND-22 : 5.880758445267087 ps
+NAND-23 : 5.880758445267087 ps
+
+----------------------------------------------------------------------
+	 	 CRITICAL PATH
+----------------------------------------------------------------------
+INPUT-3
+NAND-11
+NAND-16
+NAND-22
