+++ CODE/PY_FILES/Phase-2/requirements.txt
@@ -1,0 +1,3 @@
+debugpy==1.7.0
+numpy==1.21.6
+tabulate==0.9.0
